requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
pydantic
duckduckgo-search
python-multipart
